ADD: skeleton refactor to TL->CTC
